+++ clients/web3-wallet/src/binance_sdk_web3_wallet/metadata.py
@@ -1,1 +1,1 @@
-NAME = "binance-sdk-web3-wallet"
+NAME = "binance-web3-wallet"